restructure the maven modules. move to multi modules structure
