+++ .gitignore
@@ -1,0 +1,7 @@
+.idea
+target
+*.iml
+/classes/
+/generated-sources/
+/maven-status/
+/test-classes/
