+++ demo/pom.xml
@@ -1,0 +1,52 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>com.websystique.springboot</groupId>
+        <artifactId>SpringBootWeb</artifactId>
+        <version>0.4</version>
+    </parent>
+
+    <artifactId>demo</artifactId>
+    <version>0.0.1</version>
+    <packaging>jar</packaging>
+    <name>SpringBoot-Web</name>
+
+    <properties>
+        <java.version>1.8</java.version>
+        <app.version>0.4</app.version>
+    </properties>
+
+    <dependencies>
+        <!-- Add typical dependencies for a web application -->
+        <!-- Adds Tomcat and Spring MVC, along others -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-freemarker</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>com.websystique.springboot</groupId>
+            <artifactId>account-service</artifactId>
+        </dependency>
+    </dependencies>
+    <build>
+        <finalName>SpringBootWeb-${app.version}</finalName>
+        <plugins>
+            <plugin><!-- Include if you want to make an executable jar[FAT JAR which
+					includes all dependencies along with sprinboot loader] that you can run on
+					command line using java -jar NAME -->
+                <!-- If you are not including it, result would be just a regular jar that you
+                could run on command line using mvn spring-boot:run [on project root directory]-->
+                <groupId>org.springframework.boot</groupId>
+                <artifactId>spring-boot-maven-plugin</artifactId>
+                <configuration>
+                    <outputDirectory>../target</outputDirectory>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
