+++ demo/src/main/resources/application.yml
@@ -1,0 +1,3 @@
+server:
+  port: 8080
+  contextPath: /
