+++ demo/src/main/resources/templates/error.ftl
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+
+<html lang="en">
+<head>
+    <link rel="stylesheet" type="text/css" href="css/bootstrap.css" />
+</head>
+<body>
+	<div class="container">
+	    <div class="jumbotron alert-danger">
+			<h1>Oops. Something went wrong</h1>
+			<h2>${status} ${error}</h2>
+	    </div>
+	</div>
+</body>
+
+</html>
