+++ demo/src/main/resources/templates/hello.ftl
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+
+<html lang="en">
+<head>
+    <link rel="stylesheet" type="text/css" href="css/bootstrap.css" />
+</head>
+<body>
+	<div class="container">
+	    <div class="jumbotron">
+	        <h2>${title}</h2>
+	        <p>${message}</p>
+	    </div>
+	</div>
+</body>
+</html>
