--- lib/com/websystique/springboot/account-service/maven-metadata.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<metadata>
-  <groupId>com.websystique.springboot</groupId>
-  <artifactId>account-service</artifactId>
-  <versioning>
-    <release>0.0.1</release>
-    <versions>
-      <version>0.0.1</version>
-    </versions>
-    <lastUpdated>20190914193238</lastUpdated>
-  </versioning>
-</metadata>
