--- lib/com/websystique/springboot/account-service/maven-metadata.xml.sha1
@@ -1,1 +1,0 @@
-10ecfb86b27ceda2a0f7fe1de348caca5bd57708
