+++ pom.xml
@@ -1,59 +1,40 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.springframework.boot</groupId>
+        <artifactId>spring-boot-starter-parent</artifactId>
+        <version>1.4.3.RELEASE</version>
+    </parent>
+    <groupId>com.websystique.springboot</groupId>
+    <artifactId>SpringBootWeb</artifactId>
+    <version>${app.version}</version>
+    <name>SpringBoot-Web</name>
+    <packaging>pom</packaging>
 
-	<groupId>com.websystique.springboot</groupId>
-	<artifactId>SpringBootWeb</artifactId>
-	<version>0.4</version>
-	<packaging>jar</packaging>
+    <properties>
+        <java.version>1.8</java.version>
+        <app.version>0.4</app.version>
+    </properties>
 
-	<name>SpringBoot-Web</name>
+    <modules>
+        <module>demo</module>
+        <module>services</module>
+    </modules>
 
-	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.4.3.RELEASE</version>
-	</parent>
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>com.websystique.springboot</groupId>
+                <artifactId>account-service</artifactId>
+                <version>0.0.1</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 
-	<properties>
-		<java.version>1.8</java.version>
-		<!-- Additionally, Please make sure that your JAVA_HOME is pointing to 
-			1.8 when building on commandline -->
-	</properties>
-    <repositories>
-        <repository>
-            <id>localrep</id>
-            <name>local repository</name>
-            <url>file:${project.basedir}/lib</url>
-        </repository>
-    </repositories>
-	<dependencies>
-		<!-- Add typical dependencies for a web application -->
-		<!-- Adds Tomcat and Spring MVC, along others -->
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-web</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-freemarker</artifactId>
-		</dependency>
-        <dependency>
-            <groupId>com.websystique.springboot</groupId>
-            <artifactId>account-service</artifactId>
-            <version>0.0.1</version>
-        </dependency>
-	</dependencies>
-	<build>
-		<plugins>
-			<plugin><!-- Include if you want to make an executable jar[FAT JAR which 
-					includes all dependencies along with sprinboot loader] that you can run on 
-					command line using java -jar NAME -->
-				<!-- If you are not including it, result would be just a regular jar that you 
-				could run on command line using mvn spring-boot:run [on project root directory]-->
-				<groupId>org.springframework.boot</groupId>
-				<artifactId>spring-boot-maven-plugin</artifactId>
-			</plugin>
-		</plugins>
-	</build>
+    <build>
+        <plugins>
+
+        </plugins>
+    </build>
 </project>
