+++ services/account-service/pom.xml
@@ -45,10 +45,8 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>lib/</classpathPrefix>
-<!--                            <mainClass>ru.dokwork.App</mainClass>-->
                         </manifest>
                     </archive>
-                    <outputDirectory>../../lib</outputDirectory>
                 </configuration>
             </plugin>
         </plugins>
