+++ services/pom.xml
@@ -2,11 +2,10 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-starter-parent</artifactId>
-        <version>1.4.3.RELEASE</version>
+        <groupId>com.websystique.springboot</groupId>
+        <artifactId>SpringBootWeb</artifactId>
+        <version>0.4</version>
     </parent>
-    <groupId>com.websystique.springboot</groupId>
     <artifactId>services</artifactId>
     <version>0.0.1</version>
     <packaging>pom</packaging>
