--- src/main/java/com/websystique/springboot/SpringBootStandAloneApp.java
@@ -1,20 +1,0 @@
-package com.websystique.springboot;
-
-import com.websystique.micro.service.account.domain.Account;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.context.annotation.*;
-
-@EnableAutoConfiguration
-@ComponentScans(
-        {
-                @ComponentScan(value = "com.websystique.springboot"),
-                @ComponentScan(value = "com.websystique.micro.service.account")
-        })
-//@SpringBootApplication(scanBasePackages={"com.websystique.springboot"})// same as @Configuration @EnableAutoConfiguration @ComponentScan
-public class SpringBootStandAloneApp {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringBootStandAloneApp.class, args);
-    }
-}
