--- src/main/resources/application.yml
@@ -1,3 +1,0 @@
-server:
-  port: 8080
-  contextPath: /
